Move Player class from demos to core
