--- music/demos/tone.d.ts
@@ -1,3 +1,0 @@
-// Module declaration for Tone.js, making it explicitly instead of implicitly
-// any for the TypeScript compiler.
-declare module "tone";
