+++ music/src/core/tone.d.ts
@@ -1,0 +1,1 @@
+declare module 'tone';
